+++ src/classes/Application/Admin/Area/Mode/CollectionList.php
@@ -1,6 +1,10 @@
 <?php
 
-abstract class Application_Admin_Area_Mode_CollectionList extends Application_Admin_Area_Mode
+use Application\Interfaces\Admin\CollectionListInterface;
+
+abstract class Application_Admin_Area_Mode_CollectionList
+    extends Application_Admin_Area_Mode
+    implements CollectionListInterface
 {
     use Application_Traits_Admin_CollectionList;